refactor: consolidate BASE_URL and other constants into lib/constants and update imports
