+++ app/blog/[slug]/page.tsx
@@ -1,9 +1,9 @@
 import { notFound } from "next/navigation";
 import { CustomMDX } from "@/components/composite/mdx";
 import { formatDate, getBlogPosts } from "@/lib/blog/utils";
-import { BASE_URL } from "app/sitemap";
 import { BadgeLink } from "@/components/composite/badge-link";
 import { FaArrowLeft } from "react-icons/fa6";
+import { BASE_URL } from "@/lib/constants";
 
 export async function generateStaticParams() {
   const posts = getBlogPosts();
