+++ lib/constants.ts
@@ -1,9 +1,12 @@
+import { z } from "zod";
+
 export const TITLE = "Zach's Portfolio";
 export const DESCRIPTION =
   "A minimal blog and landing page to get connected with me.";
-export const SITE_LINKS = {
+export const RELATIVE_SITE_LINKS = {
   HOME: "/",
   BLOG: "/blog",
+  BLOG_POST: (slug: string) => `/blog/${slug}`,
 } as const;
 
 export const EXTERNAL_LINKS = {
@@ -15,3 +18,32 @@ export const EXTERNAL_LINKS = {
   BLUESKY: "https://bsky.app/profile/zacowan.bsky.social",
   MASTODON: "https://mastodon.social/@zacowan",
 } as const;
+
+// https://vercel.com/docs/projects/environment-variables/system-environment-variables#system-environment-variables
+const vercelEnvironmentSchema = z.object({
+  VERCEL_ENV: z
+    .literal("production")
+    .or(z.literal("preview"))
+    .or(z.literal("development")),
+  VERCEL_PROJECT_PRODUCTION_URL: z.string(),
+  VERCEL_BRANCH_URL: z.string(),
+  VERCEL_URL: z.string(),
+});
+
+const LOCAL_URL = "https://localhost:3000";
+
+const getBaseUrl = () => {
+  try {
+    const safeEnv = vercelEnvironmentSchema.parse(process.env);
+    const url =
+      safeEnv.VERCEL_ENV === "production"
+        ? `https://${safeEnv.VERCEL_PROJECT_PRODUCTION_URL}`
+        : `https://${safeEnv.VERCEL_URL}`;
+    console.log(`Vercel environment variables found, using ${url} as base url`);
+    return url;
+  } catch {
+    return LOCAL_URL;
+  }
+};
+
+export const BASE_URL = getBaseUrl();
